# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PLMKR \u2014 Playmaker API",
    "description": "AI-powered artist management platform. Agents handle curator pitching, PR outreach, venue booking, social scheduling, and weekly reporting \u2014 no human copy-paste required.\n\n**Phases:**\n- Phase 1: Gmail OAuth + Curator pitch lifecycle\n- Phase 2: PR contacts + Booking inquiries\n- Phase 3: Social post scheduling + Weekly reports\n\nAll endpoints require an `artist_id` (the artist's unique identifier in the system).",
    "contact": {
      "name": "PLMKR Support",
      "email": "getnexusai@gmail.com"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "3.0.0"
  },
  "paths": {
...
    "/api/admin/stats": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Admin Stats",
        "description": "Return activity stats for an artist.\nsince: ISO datetime string (optional). If omitted, returns all-time stats.",
        "operationId": "admin_stats_api_admin_stats_get",
        "parameters": [
          {
            "name": "artist_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artist Id"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}